img: use variables density and preservation

+++ img-compress.sh
@@ -14,10 +14,6 @@ maxwidth=1750 # at 300dpi = 148mm (A4 main body width, circa)
 lowheight=500
 lowwidth=800
 
-# add safe max values for HTML publication (eg: 800 width, a good guess for height?)
-# exclusion of PNG if named differently (how?)
-# 2 versions per image
-# print width and height of each original image (if lower than "maxwidth"?)
 
 # check if the folder for original files is there
 if [ ! -d orig ]; then
@@ -66,6 +62,7 @@ else
 			-d|--dpi)
 				if [ "$2" -eq "$2" ] 2>/dev/null
 				then
+					# set a different density
 					DENSITY="$2"
 				else
 					echo "ERROR: --dpi must be an integer."
@@ -92,25 +89,32 @@ for image in *.{jpeg,jpg,png,tiff} ; do
 	cp ${image} ./orig/${image}
 
 	if [ "${image}" != "${image%.${EXTPNG}}" ]; then
-		convert ${image} -resize ${lowwidth}x${lowheight}\> ${image%.${EXTPNG}}.low.jpg
-		convert -units PixelsPerInch ${image} -density 300 ${image%.${EXTPNG}}.jpg
-		convert ${image%.${EXTPNG}}.jpg -resize ${maxwidth}x${maxheight}\> ${image%.${EXTPNG}}.jpg
-		rm ${image}
-		echo "${image} converted in JPG and resized (if necessary)"
+		# if "preserve"
+		if [ $p ]; then
+			convert -units PixelsPerInch ${image} -density $DENSITY ${image%.${EXTPNG}}.png
+			convert ${image%.${EXTPNG}}.png -resize ${maxwidth}x${maxheight}\> ${image%.${EXTPNG}}.png
+			echo "${image} converted keeped as PNG and resized (if necessary)"
+		else
+			convert ${image} -resize ${lowwidth}x${lowheight}\> ${image%.${EXTPNG}}.low.jpg
+			convert -units PixelsPerInch ${image} -density $DENSITY ${image%.${EXTPNG}}.jpg
+			convert ${image%.${EXTPNG}}.jpg -resize ${maxwidth}x${maxheight}\> ${image%.${EXTPNG}}.jpg
+			rm ${image}
+			echo "${image} converted in JPG and resized (if necessary)"
+		fi
 	elif [ "${image}" != "${image%.${EXTJPEG}}" ]; then
 		convert ${image} -resize ${lowwidth}x${lowheight}\> ${image%.${EXTJPEG}}.low.jpg
-		convert -units PixelsPerInch ${image} -density 300 ${image%.${EXTJPEG}}.jpg
+		convert -units PixelsPerInch ${image} -density $DENSITY ${image%.${EXTJPEG}}.jpg
 		convert ${image%.${EXTJPEG}}.jpg -resize ${maxwidth}x${maxheight}\> ${image%.${EXTJPEG}}.jpg
 		rm ${image}
 		echo "${image} renamed in JPG and resized (if necessary)"
 	elif [ "${image}" != "${image%.${EXTJPG}}" ]; then
 		convert ${image} -resize ${lowwidth}x${lowheight}\> ${image%.${EXTJPG}}.low.jpg
-		convert -units PixelsPerInch ${image} -density 300 ${image%.${EXTJPG}}.jpg
+		convert -units PixelsPerInch ${image} -density $DENSITY ${image%.${EXTJPG}}.jpg
 		convert ${image%.${EXTJPG}}.jpg -resize ${maxwidth}x${maxheight}\> ${image%.${EXTJPG}}.jpg
 		echo "${image} resized (if necessary)"
 	elif [ "${image}" != "${image%.${EXTTIFF}}" ]; then
 		convert ${image} -resize ${lowwidth}x${lowheight}\> ${image%.${EXTTIFF}}.low.jpg
-		convert -units PixelsPerInch ${image} -density 300 ${image%.${EXTTIFF}}.jpg
+		convert -units PixelsPerInch ${image} -density $DENSITY ${image%.${EXTTIFF}}.jpg
 		convert ${image%.${EXTTIFF}}.jpg -resize ${maxwidth}x${maxheight}\> ${image%.${EXTTIFF}}.jpg
 		rm ${image}
 		echo "${image} converted in JPG and resized (if necessary)"
